requirements.txt
pandas>=1.5.0
numpy>=1.23.0
pyyaml>=6.0
matplotlib>=3.5.0
requests>=2.28.0
python-dotenv>=1.0.0

